refactor: isolate helper api functions
